+++ internal/examples/sycl/allreduce_sycl.cpp
@@ -18,7 +18,7 @@
 
 template <class processing_type>
 void user_thread_idx(size_t thread_idx,
-                     const std::vector<std::pair<size_t, cl::sycl::device>>& devices,
+                     const std::vector<std::pair<int, cl::sycl::device>>& devices,
                      cl::sycl::context ctx,
                      int total_devices_in_cluster,
                      std::shared_ptr<ccl::kvs_interface> kvs_instance,
@@ -242,7 +242,7 @@ int main(int argc, char** argv) {
         std::vector<device_type> devices;
         std::tie(thread_id, devices) = *thread_affinity_it;
 
-        std::vector<std::pair<size_t, device_type>> ranked_devices;
+        std::vector<std::pair<int, device_type>> ranked_devices;
         ranked_devices.reserve(devices.size());
         std::transform(devices.begin(),
                        devices.end(),
